refactored validators namespace and added support for validation errors as request result error messages
