+++ Apps/Common/src/ErrorHandling/RequestResultFactory.cs
@@ -16,6 +16,8 @@
 
 namespace HealthGateway.Common.ErrorHandling
 {
+    using System.Collections.Generic;
+    using System.Linq;
     using HealthGateway.Common.Data.ViewModels;
 
     /// <summary>
@@ -27,18 +29,40 @@ namespace HealthGateway.Common.ErrorHandling
         /// Factory method for error <see cref="RequestResult{T}"/> instances.
         /// </summary>
         /// <param name="errorType">The error type.</param>
-        /// <param name="errorMessages">The error messages.</param>
+        /// <param name="errorMessage">The error message.</param>
         /// <typeparam name="T">The payload type.</typeparam>
         /// <returns>New  <see cref="RequestResult{T}"/> instance with error.</returns>
-        public static RequestResult<T> Error<T>(ErrorType errorType, params string[] errorMessages)
+        public static RequestResult<T> Error<T>(ErrorType errorType, string errorMessage)
             where T : class? => new RequestResult<T>
             {
                 ResultStatus = Data.Constants.ResultType.Error,
                 ResultError = new RequestResultError
                 {
-                    ResultMessage = string.Join(";", errorMessages),
+                    ResultMessage = errorMessage,
                     ErrorCode = ErrorTranslator.InternalError(errorType),
                 },
             };
+
+        /// <summary>
+        /// Factory method for error <see cref="RequestResult{T}"/> instances.
+        /// </summary>
+        /// <param name="errorType">The error type.</param>
+        /// <param name="errorMessages">The error messages.</param>
+        /// <typeparam name="T">The payload type.</typeparam>
+        /// <returns>New  <see cref="RequestResult{T}"/> instance with error.</returns>
+        public static RequestResult<T> Error<T>(ErrorType errorType, IEnumerable<string> errorMessages)
+            where T : class? =>
+                Error<T>(errorType, string.Join(";", errorMessages));
+
+        /// <summary>
+        /// Factory method for error <see cref="RequestResult{T}"/> instances.
+        /// </summary>
+        /// <param name="errorType">The error type.</param>
+        /// <param name="validationResults">Fluent validation errors.</param>
+        /// <typeparam name="T">The payload type.</typeparam>
+        /// <returns>New  <see cref="RequestResult{T}"/> instance with error.</returns>
+        public static RequestResult<T> Error<T>(ErrorType errorType, IEnumerable<FluentValidation.Results.ValidationFailure> validationResults)
+            where T : class? =>
+                Error<T>(errorType, validationResults.Select(vr => vr.ErrorMessage));
     }
 }
