+++ Apps/GatewayApi/src/Validations/AddDependentRequestValidator.cs
@@ -14,23 +14,25 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------
 
-namespace HealthGateway.Common.Validations
+namespace HealthGateway.GatewayApi.Validations
 {
     using FluentValidation;
     using HealthGateway.Common.Data.Validations;
-    using HealthGateway.Common.Models.PHSA;
+    using HealthGateway.GatewayApi.Models;
 
     /// <summary>
-    /// Validation rules for <see cref="VaccineStatusQuery"/>.
+    /// Validates <see cref="AddDependentRequestValidator"/> instances.
     /// </summary>
-    public class VaccineStatusQueryValidator : AbstractValidator<VaccineStatusQuery>
+    public class AddDependentRequestValidator : AbstractValidator<AddDependentRequest>
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="VaccineStatusQueryValidator"/> class.
+        /// Initializes a new instance of the <see cref="AddDependentRequestValidator"/> class.
         /// </summary>
-        public VaccineStatusQueryValidator()
+        /// <param name="maxDependantAge">optionally set the maximum age of a dependent, defaults to 12.</param>
+        public AddDependentRequestValidator(int maxDependantAge = 12)
         {
-            this.RuleFor(v => v.PersonalHealthNumber).NotEmpty().SetValidator(new PhnValidator());
+            this.RuleFor(v => v.Phn).SetValidator(new PhnValidator());
+            this.RuleFor(v => v.DateOfBirth).SetValidator(new DependantAgeValidator(maxDependentAge: maxDependantAge));
         }
     }
 }
