+++ Apps/Immunization/src/Services/VaccineStatusService.cs
@@ -16,7 +16,6 @@
 namespace HealthGateway.Immunization.Services
 {
     using System;
-    using System.Linq;
     using System.Threading.Tasks;
     using AutoMapper;
     using HealthGateway.Common.AccessManagement.Authentication;
@@ -207,7 +206,7 @@ namespace HealthGateway.Immunization.Services
             var validationResults = new VaccineStatusQueryValidator().Validate(query);
             if (!validationResults.IsValid)
             {
-                return RequestResultFactory.Error<VaccineStatus>(ErrorType.InvalidState, validationResults.Errors.Select(e => e.ErrorMessage).ToArray());
+                return RequestResultFactory.Error<VaccineStatus>(ErrorType.InvalidState, validationResults.Errors);
             }
 
             string? accessToken = this.authDelegate.AuthenticateAsSystem(this.tokenUri, this.tokenRequest).AccessToken;
